requirements.txt
certifi~=2019.11.28
charset-normalizer==2.0.6
idna~=2.8
loguru==0.5.3
pyTelegramBotAPI==4.0.1
python-dotenv==0.19.0
requests~=2.22.0
urllib3~=1.25.8
pip~=20.0.2
keyring~=18.0.1
Click~=7.0
Pillow~=7.0.0
cryptography~=2.8
six~=1.14.0
chardet~=3.0.4
simplejson~=3.16.0
setuptools~=45.2.0